requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.17.0
numpy>=1.24.0 